requirements.txt
numpy==1.23.5
torch==1.12.1
huggingface-hub==0.20.3
transformers==4.30.0
diffusers==0.21.0
pillow==10.0.0
scikit-image==0.21.0
accelerate==0.21.0
pydantic==1.10.13
gradio==3.50.0
matplotlib==3.7.0
pyyaml==6.0.1
sqlalchemy==2.0.20
psutil==5.9.5
